# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Frameless Creative Admin API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /filmmaking-documents:
    get:
      operationId: listFilmmakingDocuments
      tags: [filmmaking-documents]
      summary: List crew's filmmaking documents
      parameters:
        - in: query
          name: projectId
          schema:
            type: string
          description: Optional project ID filter
      responses:
        "200":
          description: List of documents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilmmakingDocument"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createFilmmakingDocument
      tags: [filmmaking-documents]
      summary: Create new filmmaking document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFilmmakingDocumentBody"
      responses:
        "201":
          description: Document created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingDocument"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-documents/{id}:
    get:
      operationId: getFilmmakingDocument
      tags: [filmmaking-documents]
      summary: Get filmmaking document by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingDocumentDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateFilmmakingDocument
      tags: [filmmaking-documents]
      summary: Update filmmaking document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateFilmmakingDocumentBody"
      responses:
        "200":
          description: Document updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingDocument"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteFilmmakingDocument
      tags: [filmmaking-documents]
      summary: Delete filmmaking document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Document deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-documents/{id}/duplicate:
    post:
      operationId: duplicateFilmmakingDocument
      tags: [filmmaking-documents]
      summary: Clone/duplicate document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "201":
          description: Document duplicated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingDocument"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-documents/{id}/collaborators:
    get:
      operationId: listDocumentCollaborators
      tags: [filmmaking-collaborators]
      summary: List document collaborators
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: List of collaborators
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FilmmakingCollaboratorDetail"
    post:
      operationId: addDocumentCollaborator
      tags: [filmmaking-collaborators]
      summary: Add collaborator to document
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCollaboratorBody"
      responses:
        "201":
          description: Collaborator added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingCollaborator"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /filmmaking-documents/{id}/collaborators/{collaboratorId}:
    patch:
      operationId: updateDocumentCollaborator
      tags: [filmmaking-collaborators]
      summary: Update collaborator role
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: collaboratorId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCollaboratorBody"
      responses:
        "200":
          description: Collaborator updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FilmmakingCollaborator"
    delete:
      operationId: removeDocumentCollaborator
      tags: [filmmaking-collaborators]
      summary: Remove collaborator
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
        - in: path
          name: collaboratorId
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Collaborator removed
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    FilmmakingDocument:
      type: object
      properties:
        id:
          type: string
        projectId:
          type: string
          nullable: true
        crewId:
          type: string
        docType:
          type: string
          enum: [concept, script, shotlist]
        title:
          type: string
        content:
          type: object
          nullable: true
        isDraft:
          type: boolean
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, crewId, docType, title, isDraft, createdAt, updatedAt]
    FilmmakingDocumentDetail:
      allOf:
        - $ref: "#/components/schemas/FilmmakingDocument"
        - type: object
          properties:
            collaborators:
              type: array
              items:
                $ref: "#/components/schemas/FilmmakingCollaborator"
            userAccess:
              type: string
              enum: [owner, editor, viewer]
          required: [collaborators, userAccess]
    CreateFilmmakingDocumentBody:
      type: object
      properties:
        docType:
          type: string
          enum: [concept, script, shotlist]
        title:
          type: string
        projectId:
          type: string
          nullable: true
        content:
          type: object
          nullable: true
      required: [docType, title]
    UpdateFilmmakingDocumentBody:
      type: object
      properties:
        content:
          type: object
          nullable: true
        title:
          type: string
          nullable: true
        isDraft:
          type: boolean
        changeSummary:
          type: string
          nullable: true
    FilmmakingCollaborator:
      type: object
      properties:
        id:
          type: string
        documentId:
          type: string
        crewMemberId:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]
        addedBy:
          type: string
        addedAt:
          type: string
      required: [id, documentId, crewMemberId, role, addedBy, addedAt]
    FilmmakingCollaboratorDetail:
      allOf:
        - $ref: "#/components/schemas/FilmmakingCollaborator"
        - type: object
          properties:
            memberName:
              type: string
              nullable: true
            memberEmail:
              type: string
              nullable: true
            memberAvatarUrl:
              type: string
              nullable: true
    AddCollaboratorBody:
      type: object
      properties:
        crewMemberId:
          type: string
        role:
          type: string
          enum: [editor, viewer]
      required: [crewMemberId, role]
    UpdateCollaboratorBody:
      type: object
      properties:
        role:
          type: string
          enum: [editor, viewer]
      required: [role]
...